requirements.txt
asgiref==3.8.1
Django==5.1.3
django-cors-headers==4.6.0
django-filter==24.3
django-safedelete==1.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
packaging==24.2
PyJWT==2.9.0
pytz==2024.2
sqlparse==0.5.1
